StartTime null check

+++ _ChocoClean.ps1
@@ -1,2 +1,5 @@
-$Desktops = "$env:PUBLIC\Desktop", "$env:USERPROFILE\Desktop"
-$Desktops | Get-ChildItem -Filter "*.lnk" -ErrorAction SilentlyContinue | Where-Object { $_.LastWriteTime -gt $StartTime } | Remove-Item
+If ($StartTime -ne $null)
+{
+    $Desktops = "$env:PUBLIC\Desktop", "$env:USERPROFILE\Desktop"
+    $Desktops | Get-ChildItem -Filter "*.lnk" -ErrorAction SilentlyContinue | Where-Object { $_.LastWriteTime -gt $StartTime } | Remove-Item
+}
